chore: association field default value
